pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api-governor"
version = "1.0.0"
description = "API governance and breaking change detection for OpenAPI specs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "API Governance Skill Contributors" }
]
keywords = [
    "openapi",
    "api",
    "governance",
    "linting",
    "breaking-changes",
    "validation",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "types-PyYAML",
]

[project.scripts]
api-governor = "api_governor.__main__:main"

[project.urls]
Homepage = "https://github.com/yourorg/api-governance-skill"
Documentation = "https://github.com/yourorg/api-governance-skill#readme"
Repository = "https://github.com/yourorg/api-governance-skill"
Issues = "https://github.com/yourorg/api-governance-skill/issues"
Changelog = "https://github.com/yourorg/api-governance-skill/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/skills",
    "/schemas",
]

[tool.hatch.build.targets.wheel]
packages = ["src/api_governor"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["api_governor"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "yaml.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/api_governor"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
